Fixed build script + updated readme

+++ build.ps1
@@ -20,15 +20,27 @@ Add-Content $HELP_FILE "- DO NOT UNZIP THE TWO EXTRACTED FILES"
 Add-Content $HELP_FILE "- To install the datapack: drag and drop the datapack file inside the Data Packs screen when creating a new world, enable it and click 'Done'"
 Add-Content $HELP_FILE "- To install the resourcepack: drag and drop the resourcepack file inside the Resource Packs screen or folder, enable it and click 'Done'"
 
+#Setup zip files
+Add-Type -AssemblyName System.IO.Compression
+Add-Type -AssemblyName System.IO.Compression.FileSystem
+
 #Create resource pack
-Compress-Archive -Path $(".\"+$RESOURCE_PACK_FOLDER+"\*"), $PACK_PNG_PATH -DestinationPath $RESOURCE_PACK_PATH
+if (Test-Path $RESOURCE_PACK_PATH) {
+    Remove-Item $RESOURCE_PACK_PATH -Force
+}
+7z a $RESOURCE_PACK_PATH $(".\"+$RESOURCE_PACK_FOLDER+"\*") $PACK_PNG_PATH
 
 #Create data pack
-Compress-Archive -Path $(".\"+$DATA_PACK_FOLDER+"\*"), $PACK_PNG_PATH -DestinationPath $DATA_PACK_PATH
+if (Test-Path $DATA_PACK_PATH) {
+    Remove-Item $DATA_PACK_PATH -Force
+}
+7z a $DATA_PACK_PATH $(".\"+$DATA_PACK_FOLDER+"\*") $PACK_PNG_PATH
 
 #Create Full pack file
-Remove-Item $FULL_PACK_PATH -Force
-Compress-Archive -Path $RESOURCE_PACK_PATH, $DATA_PACK_PATH, $HELP_FILE -DestinationPath $FULL_PACK_PATH
+if (Test-Path $FULL_PACK_PATH) {
+    Remove-Item $FULL_PACK_PATH -Force
+}
+7z a $FULL_PACK_PATH $RESOURCE_PACK_PATH $DATA_PACK_PATH $HELP_FILE
 
 #Delete resource pack file
 Remove-Item $RESOURCE_PACK_PATH -Force
